Many modifications
The main modification is the use of a much more sophisticated tokenizer, which is now based on a true automaton implementation. We also introduce some mechanism such as tail recursion, which can be inferred or forced in some cases.

This tokenizer works with most encodings: string, u32string, u16string etc.
However, it is implemented on the top of u32string.

+++ include/equationtemplate.h
@@ -2531,7 +2531,7 @@ public:
             vector<Tamgu*> vv;
             for (long i = 0; i < v.size(); i++) {
                 if (v[i]->isConst())
-                    val += v[i]->Long();
+                    val += v[i]->Float();
                 else
                     vv.push_back(v[i]);
             }
@@ -2883,13 +2883,13 @@ public:
                 if (v[i]->isConst()) {
                     if (!i) {
                         first = true;
-                        val = v[i]->Long();
+                        val = v[i]->Float();
                         continue;
                     }
                     if (first)
-                        val -= v[i]->Long();
+                        val -= v[i]->Float();
                     else
-                        val += v[i]->Long();
+                        val += v[i]->Float();
                 }
                 else
                     vv.push_back(v[i]);
@@ -3144,7 +3144,7 @@ public:
             vector<Tamgu*> vv;
             for (long i = 0; i < v.size(); i++) {
                 if (v[i]->isConst())
-                    val *= v[i]->Long();
+                    val *= v[i]->Float();
                 else
                     vv.push_back(v[i]);
             }
